Move TaskStatus enum into `grakn-core`
