+++ grakn-core/src/main/java/ai/grakn/engine/TaskStatus.java
@@ -16,7 +16,7 @@
  * along with Grakn. If not, see <http://www.gnu.org/licenses/gpl.txt>.
  */
 
-package ai.grakn.engine.backgroundtasks;
+package ai.grakn.engine;
 
 import java.io.Serializable;
