+++ grakn-test/src/test/java/ai/grakn/test/engine/backgroundtasks/BackgroundTaskTestUtils.java
@@ -20,7 +20,7 @@ package ai.grakn.test.engine.backgroundtasks;
 
 import ai.grakn.engine.backgroundtasks.TaskState;
 import ai.grakn.engine.backgroundtasks.TaskStateStorage;
-import ai.grakn.engine.backgroundtasks.TaskStatus;
+import ai.grakn.engine.TaskStatus;
 import mjson.Json;
 
 import java.util.Date;
