+++ grakn-test/src/test/java/ai/grakn/test/engine/backgroundtasks/TaskStateZookeeperStoreTest.java
@@ -32,8 +32,8 @@ import java.util.Set;
 import java.util.UUID;
 import java.util.stream.Collectors;
 
-import static ai.grakn.engine.backgroundtasks.TaskStatus.CREATED;
-import static ai.grakn.engine.backgroundtasks.TaskStatus.SCHEDULED;
+import static ai.grakn.engine.TaskStatus.CREATED;
+import static ai.grakn.engine.TaskStatus.SCHEDULED;
 import static java.time.Instant.now;
 import static junit.framework.TestCase.assertEquals;
 import static org.junit.Assert.assertFalse;
